# backend/docs/swagger.yaml
basePath: /admin-api
definitions:
  github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      msg:
        example: ""
        type: string
    type: object
...
  internal_modules_pay.AppSaveRequest:
    properties:
      appKey:
        type: string
      id:
        type: integer
      name:
        type: string
      orderNotifyUrl:
        type: string
      refundNotifyUrl:
        type: string
      remark:
        type: string
      status:
        type: integer
      transferNotifyUrl:
        type: string
    required:
    - name
    type: object
  internal_modules_pay.ChannelSaveRequest:
    properties:
      appId:
        type: integer
      code:
        type: string
      config:
        type: string
      feeRate:
        type: integer
      id:
        type: integer
      remark:
        type: string
      status:
        type: integer
    required:
    - appId
    - code
    type: object
  internal_modules_pay.OrderCreateRequest:
    properties:
      appId:
        type: integer
      body:
        type: string
      channelCode:
        type: string
      merchantOrderId:
        type: string
      price:
        type: integer
      subject:
        type: string
    required:
    - appId
    - merchantOrderId
    - price
    - subject
    type: object
  internal_modules_pay.RefundCreateRequest:
    properties:
      merchantRefundId:
        type: string
      orderId:
        type: integer
      reason:
        type: string
      refundPrice:
        type: integer
    required:
    - merchantRefundId
    - orderId
    - refundPrice
    type: object
...
info:
  contact: {}
  description: 'Go modular-monolith scaffold for operations platforms. Default database:
    MySQL 8.4.'
  title: Nimbus Framework Go API
  version: "1.0"
paths:
...
  /pay/app/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payment app
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_pay.AppSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Create a payment application
      tags:
      - Pay App
  /pay/app/delete:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete a payment application
      tags:
      - Pay App
  /pay/app/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Get a payment application
      tags:
      - Pay App
  /pay/app/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List payment applications
      tags:
      - Pay App
  /pay/app/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page payment applications
      tags:
      - Pay App
  /pay/app/simple-list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List enabled payment applications
      tags:
      - Pay App
  /pay/app/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Payment app
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_pay.AppSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Update a payment application
      tags:
      - Pay App
  /pay/app/update-status:
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Change a payment application status
      tags:
      - Pay App
  /pay/channel/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payment channel
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_pay.ChannelSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Create a payment channel
      tags:
      - Pay Channel
  /pay/channel/delete:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete a payment channel
      tags:
      - Pay Channel
...
  /pay/channel/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Get a payment channel
      tags:
      - Pay Channel
  /pay/channel/get-enable-channel-code-list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List enabled channel codes for an application
      tags:
      - Pay Channel
  /pay/channel/get-enable-code-list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List enabled channel codes for an application
      tags:
      - Pay Channel
  /pay/channel/list:
    get:
      parameters:
      - description: Payment application ID
        in: query
        name: appId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List payment channels for an application
      tags:
      - Pay Channel
  /pay/channel/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page payment channels
      tags:
      - Pay Channel
  /pay/channel/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Payment channel
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_pay.ChannelSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Update a payment channel
      tags:
      - Pay Channel
...
  /pay/order/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payment order
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_pay.OrderCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Create a pending payment order
      tags:
      - Pay Order
...
  /pay/order/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Get a payment order
      tags:
      - Pay Order
  /pay/order/get-detail:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Get a payment order detail
      tags:
      - Pay Order
  /pay/order/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page payment orders
      tags:
      - Pay Order
  /pay/order/submit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payment order
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_pay.OrderCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Create a pending payment order
      tags:
      - Pay Order
  /pay/refund/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refund
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_pay.RefundCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Create a pending refund request
      tags:
      - Pay Refund
  /pay/refund/delete:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete a pending refund record
      tags:
      - Pay Refund
...
  /pay/refund/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Get a payment refund
      tags:
      - Pay Refund
  /pay/refund/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page payment refunds
      tags:
      - Pay Refund
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"